Team_18/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
geopy==2.4.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.3.2
gunicorn==22.0.0
requests==2.31.0
Pillow==10.4.0
setuptools>=68.0.0
wheel
